requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
phonenumbers==8.12.20
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
stripe==2.56.0
